cli/requirements.txt
click>=8.1.0
pydantic>=2.10.0
PyYAML>=6.0.0
rich>=13.0.0
pytest>=8.0.0
